bandaid solution to fix issues where the bot is selecting the wrong soundboard clip
we generate a clip bank so we can access which clip to use and then use the clip with the highest confidence. the problem we had is that when we generated the clip bank, to save on memory, we only generated combinations of the file names of words consecutively and in the case where multiple clips had the same combination like how "victors wii theme" and "wii theme" both would create a combination of "wii theme", we deleted those entries entirely (this was meant to tackle common phrases like "your mom"). our solution to this is to just rename files so this doesnt happen